backend/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
decorator==4.4.2
distro==1.8.0
exceptiongroup==1.1.3
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
moviepy==1.0.3
multidict==6.0.4
numpy==1.26.2
openai==0.28.0
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.0.1
proglog==0.1.10
protobuf==4.25.0
pyarrow==14.0.1
pydantic==2.5.0
pydantic_core==2.14.1
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.31.0
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
streamlit==1.28.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
yarl==1.9.2
zipp==3.17.0
blinker==1.7.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Werkzeug==3.0.1
zipp==3.17.0
